requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Transformers and Pretrained Models
transformers>=4.35.0
datasets>=2.14.0
huggingface-hub>=0.19.0
accelerate>=0.24.0

# OCR Models
easyocr>=1.7.0
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.1
imgaug>=0.4.0

# Layout Analysis
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
layoutparser>=0.3.4

# Document Processing
pdf2image>=1.16.3
pypdf2>=3.0.1
python-docx>=1.1.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Image Processing
scikit-image>=0.21.0
imageio>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Demo and UI
gradio>=4.0.0
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
wandb>=0.15.0  # For experiment tracking

# Evaluation
edit-distance>=1.0.6
jiwer>=3.0.0  # For WER/CER calculation
nltk>=3.8.1

# LLM Integration (optional, for post-correction)
openai>=1.0.0
anthropic>=0.7.0

# Utilities for handwriting-specific tasks
handwriting-ocr>=0.0.1  # If available
labelme>=5.3.0  # For annotation if needed

# Model serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
jupyterlab>=4.0.0
